Cpu read and write method maked

+++ inc/common.h
@@ -4,6 +4,6 @@
 #define COMMON_H
 
 typedef uint8_t byte;
-typedef uint8_t word;
+typedef uint16_t word;
 
 #endif
